Add small lib to allow image rendering in both site and GitHub
