--- next.config.js
@@ -1,21 +1,0 @@
-/** @type {import('next').NextConfig} */
-const path = require("path");
-const createMDX = require("@next/mdx");
-
-const nextConfig = {
-  sassOptions: {
-    includePaths: [
-      path.join(__dirname, "./", "node_modules", "@uswds", "uswds", "packages"),
-    ],
-  },
-  transpilePackages: ["yaml"],
-  output: "standalone",
-  basePath: "",
-  pageExtensions: ["js", "jsx", "md", "mdx", "ts", "tsx"],
-};
-
-const withMDX = createMDX({
-  extension: /\.(md|mdx)$/,
-});
-
-module.exports = withMDX(nextConfig);
