+++ package.json
@@ -49,7 +49,8 @@
     "react-idle-timer": "^5.7.2",
     "react-loading-skeleton": "^3.5.0",
     "react-toastify": "^10.0.5",
-    "sharp": "^0.33.5"
+    "sharp": "^0.33.5",
+    "unist-util-visit": "^5.0.0"
   },
   "devDependencies": {
     "@next/eslint-plugin-next": "^14.0.4",
